GatorTraderBackend/requirements.txt
flask
bcrypt
psycopg2
pytest
pyjwt
